Added cosine similarity testing
Be careful with this. It can be easy to reach the rate limit for OpenAi in one session, if too many tests are added. If I had more time, I would recommend that we make a function to batch requests. That seems quite overblown for a system that does not need to bear that load, however.